feat: let's go refactoring! (and to the moon!)
* hardware.pica: proper `DisplayController` typing
* mango.backend.Horizon: detect GPU Hangs and panic
* mango: some more validation
* mango: set cubemap registers (just needs testing)
* mango: refactor (needs a lot more)
* all: update to latest zig (as always)

+++ src/hardware.zig
@@ -28,7 +28,7 @@ pub fn AlignedPhysicalAddress(comptime address_alignment: std.mem.Alignment, com
         pub const alignment = address_alignment;
         pub const shift = address_shift;
 
-        pub inline fn fromAddress(address: usize) AlignedPhysAddr {
+        pub inline fn fromAddress(address: u32) AlignedPhysAddr {
             return .fromPhysical(@as(PhysicalAddress, @enumFromInt(address)));
         }
